requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
python-dotenv
pydantic
